remove redundand folder
