--- src/auth/_auth.scss
@@ -1,44 +1,0 @@
-.auth-login {
-  margin: auto;
-  width: 460px;
-
-  &__title {
-    font-size: 27.5px;
-    font-weight: bold;
-    font-style: normal;
-    font-stretch: normal;
-    line-height: 1.22;
-    letter-spacing: normal;
-    text-align: center;
-    margin: 0 0 12px 0;
-  }
-
-  &__description {
-    text-align: center;
-    color: #525a66;
-    margin: 0 0 44px;
-  }
-
-  &__providers {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-
-    li {
-      margin: 0 0 15px;
-      display: block;
-    }
-  }
-
-  &__provider {
-    border: solid 2px #525a66;
-    display: block;
-    line-height: 44px;
-    text-align: center;
-  }
-}
-
-.auth-login-form__submit {
-  text-transform: uppercase;
-  font-weight: bold;
-}
