--- src/auth/constants.js
@@ -1,1 +1,0 @@
-// @flow
