--- src/roleSelector/RoleSelector.js
@@ -1,15 +1,0 @@
-import React, {Component} from 'react';
-
-class RoleSelector extends Component {
-  static propTypes = {};
-
-  render() {
-    return (
-      <div className="section__container">
-        <p>Role-selector</p>
-      </div>
-    );
-  }
-}
-
-export default RoleSelector;
